changed my step def location
